FIX: fractional-share order fill verification crashed with uncaught ValueError
OrderManager.get_order_filled_quantity() called int(filled_qty) on Alpaca's
response, but Alpaca returns filled_qty as a string to preserve decimal
precision (e.g. "4.87"). This system actively trades fractional shares
(confirmed via real open positions with quantities like 0.50/2.81/5.87
shares), so int("4.87") raised an uncaught ValueError - not caught by the
retry loop's except clause, which only handles requests.RequestException/
Timeout. This crashed entry/exit fill verification
(executor_entry_handler.py/executor_exit_handler.py both call this function)
for any fractionally-filled live/paper order. The function's own declared
return type was already `float | None`, which int() violated.

Fixed to float(filled_qty), matching the declared type. Regression tests
confirm a fractional fill no longer raises and still returns the correct
value; confirmed via direct reproduction that int("4.87") raises before
trusting the fix (no prior test coverage existed for this function at all).

+++ algo/trading/order_manager.py
@@ -345,7 +345,14 @@ class OrderManager:
                             f"[ORDER_MANAGER] Alpaca response missing 'filled_qty' for order {alpaca_order_id}"
                         )
                         raise ValueError(f"Order {alpaca_order_id}: Alpaca response missing filled_qty (required)")
-                    return int(filled_qty)
+                    # Alpaca returns filled_qty as a STRING to preserve precision (e.g. "4.87"
+                    # for a fractional-share fill - this system actively trades fractional
+                    # shares, confirmed via real open positions). int("4.87") raises
+                    # ValueError uncaught by this function's retry loop (which only catches
+                    # requests exceptions), crashing entry/exit fill verification for any
+                    # fractionally-filled order. The function's own declared return type is
+                    # `float | None`, not int.
+                    return float(filled_qty)
                 else:
                     if attempt < max_retries - 1:
                         wait_time = 2**attempt
